[inline] shouldn't remove companion with supertype
^KT-18656 Fixed

+++ kotlin-ide/idea/src/org/jetbrains/kotlin/idea/refactoring/pullUp/KotlinPullUpHelper.kt
@@ -659,7 +659,10 @@ class KotlinPullUpHelper(
 
 internal fun KtNamedDeclaration.deleteWithCompanion() {
     val containingClass = this.containingClassOrObject
-    if (containingClass is KtObjectDeclaration && containingClass.isCompanion() && containingClass.declarations.size == 1) {
+    if (containingClass is KtObjectDeclaration &&
+        containingClass.isCompanion() &&
+        containingClass.declarations.size == 1 &&
+        containingClass.getSuperTypeList() == null) {
         containingClass.delete()
     } else {
         this.delete()
